Update script and generate new capstone proposal
The script.py and json_gen.py have been updated to reflect a change in the version from 0 to 5 and the project details. The changes include the generation of a new capstone proposal titled "Web Scraping". Additionally, new markdown and HTML files have been created for this proposal. The updated proposal now reflects Dr. Amir Jafari as the proposing personnel instead of Dr. Ryan Engstrom.