Update Project to Version 5: Time Series Model Focus
This commit updates the project to version 5, with a new focus on comparing classical and transformer-based models for time series and tabular data. Key changes include the introduction of new objective and datasets, as well as the addition of files and updates to the documentation related to this project direction.